requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
googlesearch-python==2020.0.2
idna==2.10
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.2
newspaper3k==0.2.8
nltk==3.5
Pillow==8.1.0
pkg-resources==0.0.0
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.24.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==3.1.0
tqdm==4.56.2
urllib3==1.25.9
